Add git-delta 0.4.4
A viewer for git and diff output

Code evolves, and we all spend time studying diffs. Delta aims to
make this both efficient and enjoyable: it allows you to make
extensive changes to the layout and styling of diffs, as well as
allowing you to stay arbitrarily close to the default git/diff
output.