py-django: updated to 3.1.4
Django 3.1.4 fixes several bugs in 3.1.3.

Bugfixes

Fixed setting the Content-Length HTTP header in AsyncRequestFactory.
Fixed passing extra HTTP headers to AsyncRequestFactory request methods.
Fixed crash of key transforms for JSONField on PostgreSQL when using on a Subquery() annotation.
Fixed a regression in Django 3.1 that caused a crash of auto-reloader for certain invocations of runserver on Windows with Python 3.7 and below.
Fixed a regression in Django 3.1 that caused the incorrect grouping by a Q object annotation.
Fixed a regression in Django 3.1 that caused suppressing connection errors when JSONField is used on SQLite.
Fixed a crash on SQLite, when QuerySet.values()/values_list() contained key transforms for JSONField returning non-string primitive values